quarter-2-openai-agents-sdk/class-10/openai_agents_sdk/pyproject.toml
[project]
name = "openai_agents_sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.4.2",
    "python-dotenv>=1.1.1",
]
